+++ unit_tests/environments/sugar/array.js
@@ -1431,7 +1431,19 @@ test('Array', function () {
   equal([1,2,3].sortBy(null), [1,2,3], 'Array#sortBy | null');
   equal([1,2,3].sortBy(4), [1,2,3], 'Array#sortBy | number');
 
+  var Simple = function(num) {
+    this.valueOf = function() {
+      return num;
+    }
+  }
 
+  var a = new Simple(5);
+  var b = new Simple(2);
+  var c = new Simple(3);
+  var d = new Simple(1);
+  var e = new Simple(2);
+
+  equal([a,b,c,d,e].sortBy(), [d,b,e,c,a], 'Array#sortBy | objects with "valueOf" defined will also be sorted properly');
 
   arr = [1,2,3,4,5,6,7,8,9,10];
   var firsts = [];
@@ -1676,6 +1688,463 @@ test('Array', function () {
 
   equal(arr.unique(match), arr, 'Array#unique is NOT fuzzy');
   equal([match].unique(arr), [match], 'Array#unique reverse is NOT fuzzy');
+
+
+
+  // Testing sortBy behavior
+
+  var CapturedSortOrder       = Array.AlphanumericSortOrder;
+  var CapturedSortIgnore      = Array.AlphanumericSortIgnore;
+  var CapturedSortIgnoreCase  = Array.AlphanumericSortIgnoreCase;
+  var CapturedSortEquivalents = Array.AlphanumericSortEquivalents;
+
+
+  equal([0,1,2,3,4].sortBy(), [0,1,2,3,4], 'Array#sortBy | 0 is properly sorted');
+  equal(['0','1','2','3','4'].sortBy(), ['0','1','2','3','4'], 'Array#sortBy | string numerals are properly sorted');
+  equal(['c','B','a'].sortBy(), ['a','B','c'], 'Array#sortBy | upper-case is properly sorted');
+  equal(['back','Bad','banker'].sortBy(), ['back','Bad','banker'], 'Array#sortBy | case is ignored by default');
+  equal(['c','B','a','ä','ò','p'].sortBy(), ['a','ä','B','c','ò','p'], 'Array#sortBy | should allow normalization if exists');
+  equal(['apple','apples'].sortBy(), ['apple','apples'], 'Array#sortBy | basic string length');
+  equal(['has','hàs','had','hàd'].sortBy(), ['had','hàd','has','hàs'], 'Array#sortBy | special chars basic');
+
+  arr = ['San','San Cristobal','San Juan','San Teodoro','San Tomas','Santa Barbara','Santa Clara','Santa Cruz','Santo Domingo'];
+  equal(arr.sortBy(), arr, 'Array#sortBy | spaces are counted');
+
+  equal(['AM','AB'].sortBy(), ['AB','AM'], '0 index is properly sorted');
+
+
+  arr = ['#foob','(fooc','fooa'];
+  equal(arr.sortBy(), arr, 'Array#sortBy | special chars are not ignored by default');
+
+  arr = [
+    '8braham',
+    'a4raham',
+    'abraham'
+  ];
+
+  equal(arr.sortBy(), arr, 'Array#sortBy | Numbers are filtered to the top');
+
+  arr = [
+    'pine',
+    'pino',
+    'piñata'
+  ];
+
+  equal(arr.sortBy(), arr, 'Array#sortBy | Spanish ñ is respected');
+
+  var french_names = [
+    'abelle',
+    'aceline',
+    'adélaïde',
+    'adelais',
+    'adèle',
+    'adélie',
+    'adeline',
+    'adelle',
+    'adelphe',
+    'adrienne',
+    'agace',
+    'agate',
+    'aglaë',
+    'agnès',
+    'agrippine',
+    'aimée',
+    'alaina',
+    'alais',
+    'alayna',
+    'albertine',
+    'alexandrie',
+    'alexandrine',
+    'aliénor',
+    'aline',
+    'alison',
+    'alphonsine',
+    'alvery',
+    'amaline',
+    'amandine',
+    'amarante',
+    'ambre',
+    'ambrosine',
+    'amélie',
+    'amorette',
+    'anaïs',
+    'anastaise',
+    'anastasie',
+    'andrée',
+    'andromaque',
+    'anette',
+    'angèle',
+    'angeline',
+    'angelique',
+    'ann',
+    'anne'
+  ];
+
+  equal(french_names.randomize().sortBy(), french_names, 'Array#sortBy | sorting french names');
+  equal(french_names.map('toUpperCase').randomize().sortBy(), french_names.map('toUpperCase'), 'Array#sortBy | sorting french names in upper case');
+
+
+  // MSDN http://msdn.microsoft.com/en-us/library/cc194880.aspx
+  arr = [
+    'andere',
+    'ändere',
+    'chaque',
+    'chemin',
+    'cote',
+    'cotÉ',
+    'cÔte',
+    'cÔtÉ',
+    'Czech',
+    'ČuČet',
+    'hiŠa',
+    'irdisch',
+    'lävi',
+    'lie',
+    'lire',
+    'llama',
+    'LÖwen',
+    'lÒza',
+    'LÜbeck',
+    'luck',
+    'luČ',
+    'lye',
+    'Männer',
+    'mÀŠta',
+    'mÎr',
+    'mÖchten',
+    'myndig',
+    'pint',
+    'piÑa',
+    'pylon',
+    'sämtlich',
+    'savoir',
+    'Sietla',
+    'subtle',
+    'symbol',
+    'Ślub',
+    'ŠÀran',
+    'väga',
+    'verkehrt',
+    'vox',
+    'waffle',
+    'wood',
+    'yen',
+    'yuan',
+    'yucca',
+    'zoo',
+    'ZÜrich',
+    'Zviedrija',
+    'zysk',
+    'Žal',
+    'Žena'
+  ];
+
+  equal(arr.randomize().sortBy(), arr, 'Array#sortBy | Default collation');
+
+  arr = [
+    'cweat',
+    'cwect',
+    'čweat',
+    'čweet',
+    'sweat',
+    'swect',
+    'šweat',
+    'šweet',
+    'zweat',
+    'zwect',
+    'žweat',
+    'žweet'
+  ];
+
+  equal(arr.randomize().sortBy(), arr, 'Array#sortBy | Czech/Lithuanian order is respected');
+
+
+  arr = [
+    'cat',
+    'drone',
+    'ðroll',
+    'ebert'
+  ];
+
+  equal(arr.randomize().sortBy(), arr, 'Array#sortBy | Icelandic ð order is respected');
+
+  arr = [
+    'goth',
+    'ğoad',
+    'hover',
+    'sing',
+    'şeparate',
+    'tumble'
+  ];
+
+  equal(arr.randomize().sortBy(), arr, 'Array#sortBy | Turkish order is respected');
+
+  arr = [
+    'ape',
+    'ące',
+    'central',
+    'ćenter',
+    'eulo',
+    'ęula',
+    'latch',
+    'lever',
+    'łevel',
+    'martyr',
+    'noob',
+    'ńookie',
+    'oppai',
+    'sweat',
+    'swect',
+    'śweat',
+    'śweet',
+    'yeouch',
+    'ýellow',
+    'zipper',
+    'zoophilia',
+    'źebra',
+    'żoo'
+  ];
+
+  equal(arr.randomize().sortBy(), arr, 'Array#sortBy | Polish order is respected');
+
+  arr = [
+    'cab',
+    'opec',
+    'still',
+    'zounds',
+    'æee',
+    'ølaf',
+    'ålegra'
+  ];
+
+  equal(arr.randomize().sortBy(), arr, 'Array#sortBy | Danish/Norwegian order is respected');
+
+  arr = [
+    'llama',
+    'luck',
+    'lye'
+  ];
+
+
+
+  // Compressions simply can't be handled without a complex collation system
+  // as there is simply no way fundamentally to know what was intended as a
+  // compression. For example "catch a llama" vs "catch Al Lama"
+  equal(arr.randomize().sortBy(), arr, 'Array#sortBy | Compressions are not handled');
+
+
+  arr = [
+    'àbel',
+    'abet',
+    'äpe',
+    'apu',
+    'âvec',
+    'avel',
+    'áxe',
+    'axiom',
+    'çoupon',
+    'coupos',
+    'écma',
+    'ecmo',
+    'êlam',
+    'elan',
+    'ëpic',
+    'epil',
+    'ëthen',
+    'ether',
+    'évac',
+    'eval',
+    'èxile',
+    'exilo',
+    'ïce',
+    'icy',
+    'îll',
+    'ilp',
+    'ïmpetum',
+    'impetus',
+    'íp',
+    'is',
+    'ìtalian',
+    'italians',
+    'luck',
+    'lye',
+    'òblast',
+    'oblong',
+    'ómam',
+    'omar',
+    'öpal',
+    'opam',
+    'ôva',
+    'ovum',
+    'ùla',
+    'ule',
+    'ûmar',
+    'umas',
+    'úni',
+    'uny',
+    'ùral',
+    'uranus',
+    'üte',
+    'utu'
+  ];
+
+  equal(arr.randomize().sortBy(), arr, 'Array#sortBy | Standard Western-Latin equivalents are enforced');
+
+  // Swedish collation
+  var swedish_words = [
+    'att borsta',
+    'att bränna',
+    'att brinna',
+    'att brinna',
+    'att brista',
+    'att bruka',
+    'att bryta',
+    'att bryta i bitar',
+    'att buller',
+    'att bygga',
+    'att byta',
+    'att chocka',
+    'att dela',
+    'att detaljera',
+    'att dimpa',
+    'att dö',
+    'att dö',
+    'att döda',
+    'att dofta',
+    'att dölja',
+    'att döma',
+    'att dra',
+    'att dra',
+    'att drabba',
+    'att dricka',
+    'att driva',
+    'att driva',
+    'att drömma',
+    'att duga',
+    'att erbjuda',
+    'att erkänna',
+    'att ersätta',
+    'att explodera',
+    'att falla',
+    'att falla',
+    'att fängsla',
+    'att fara',
+    'att fästa',
+    'att fastna',
+    'att fastställa',
+    'att fatta',
+    'att finna',
+    'att finna',
+    'att finnas',
+    'att fira',
+    'att fläta',
+    'att få',
+    'att fånga'
+  ];
+
+  equal(swedish_words.sortBy(), swedish_words, 'Array#sortBy | swedish strings sorted on utf8_general_ci');
+
+  var swedish_collated = [
+    'att borsta',
+    'att brinna',
+    'att brinna',
+    'att brista',
+    'att bruka',
+    'att bryta',
+    'att bryta i bitar',
+    'att bränna',
+    'att buller',
+    'att bygga',
+    'att byta',
+    'att chocka',
+    'att dela',
+    'att detaljera',
+    'att dimpa',
+    'att dofta',
+    'att dra',
+    'att dra',
+    'att drabba',
+    'att dricka',
+    'att driva',
+    'att driva',
+    'att drömma',
+    'att duga',
+    'att dö',
+    'att dö',
+    'att döda',
+    'att dölja',
+    'att döma',
+    'att erbjuda',
+    'att erkänna',
+    'att ersätta',
+    'att explodera',
+    'att falla',
+    'att falla',
+    'att fara',
+    'att fastna',
+    'att fastställa',
+    'att fatta',
+    'att finna',
+    'att finna',
+    'att finnas',
+    'att fira',
+    'att fläta',
+    'att få',
+    'att fånga',
+    'att fängsla',
+    'att fästa'
+  ];
+
+  Array.AlphanumericSortEquivalents['ö'] = null;
+  Array.AlphanumericSortEquivalents['ä'] = null;
+
+  equal(swedish_words.sortBy(), swedish_collated, 'Array#sortBy | removing equivalents can restore sort order');
+
+  // Capitals
+
+  arr = [
+    'abner',
+    'aBBey',
+    'Adrian',
+    'aDella'
+  ];
+
+  expected = [
+    'aBBey',
+    'abner',
+    'aDella',
+    'Adrian'
+  ];
+
+  Array.AlphanumericSortIgnoreCase = true;
+  equal(arr.sortBy(), expected, 'Array#sortBy | allows case ignore');
+
+
+  expected = [
+    'aDella',
+    'Adrian',
+    'aBBey',
+    'abner'
+  ];
+
+  Array.AlphanumericSortOrder = 'dba';
+  equal(arr.sortBy(), expected, 'Array#sortBy | allows other order');
+
+  expected = [
+    'aDella',
+    'abner',
+    'Adrian',
+    'aBBey'
+  ];
+
+
+  Array.AlphanumericSortIgnore = /[abcde]/g;
+  equal(arr.sortBy(), expected, 'Array#sortBy | allows custom ignore');
+
+  Array.AlphanumericSortOrder = CapturedSortOrder;
+  Array.AlphanumericSortIgnore = CapturedSortIgnore;
+  Array.AlphanumericSortIgnoreCase = CapturedSortIgnoreCase;
+  Array.AlphanumericSortEquivalents = CapturedSortEquivalents;
 
 
 });