+++ CHANGELOG.md
@@ -4,8 +4,16 @@ v1.2.3
 
 ### API Changes ###
 
+- String#compare, Number#compare, and Date#compare are deprecated.
+- Array#sortBy now makes much more sensible sorting when sorting on strings.
+- Added Array.AlphanumericSortOrder
+- Added Array.AlphanumericSortIgnore
+- Added Array.AlphanumericSortIgnoreCase
+- Added Array.AlphanumericSortEquivalents
 - Object.merge defaults are now more sensible. shallow/deep is 3rd with shallow default and resolve is 4th
 - Added Number#duration to dates package.
+- Bugfix for leaking globals.
+- Bugfix for String#compact (Issue 115)
 
 ### Internal Changes ###